add storage and ipfs components
