+++ packages/embark-ipfs/__src/upload.js
