+++ packages/embark-storage/src/index.js
@@ -8,13 +8,37 @@ class Storage {
     this.events = this.embark.events;
     this.storageConfig = embark.config.storageConfig;
     this.plugins = options.plugins;
-    let plugin = this.plugins.createPlugin('storageplugin', {});
 
-    // this.handleUploadCommand();
-    // this.addSetProviders(() => {});
+    let plugin = this.plugins.createPlugin('storageplugin', {});
     plugin.registerActionForEvent("pipeline:generateAll:before", this.addArtifactFile.bind(this));
 
-    // TODO: register xyz (e.g ipfs) to check for dappConnectionConfig
+    this.storageNodes = {};
+    this.events.setCommandHandler("storage:node:register", (clientName, startCb) => {
+      console.dir("---- registering " + clientName)
+      this.storageNodes[clientName] = startCb
+    });
+
+    this.events.setCommandHandler("storage:node:start", (storageConfig, cb) => {
+      console.dir("--------------------------");
+      console.dir("--------------------------");
+      console.dir("--------------------------");
+      console.dir("--------------------------");
+      console.dir("--------------------------");
+      console.dir("--------------------------");
+      console.dir("--- storage:node:start");
+      const clientName = storageConfig.upload.provider;
+      console.dir("---- starting...." + clientName);
+      const client = this.storageNodes[clientName];
+      if (!client) return cb("storage " + clientName + " not found");
+
+      let onStart = () => {
+        console.dir("--- storage started")
+        this.events.emit("storage:started", clientName);
+        cb();
+      }
+
+      client.apply(client, [onStart]);
+    });
   }
 
   addArtifactFile(_params, cb) {
@@ -30,64 +54,60 @@ class Storage {
     }, cb);
   }
 
-  isEnabled() {
-    return !!this.embark.config.storageConfig.enabled;
-  }
+  // handleUploadCommand() {
+  //   const self = this;
+  //   this.embark.events.setCommandHandler('storage:upload', (cb) => {
+  //     let platform = this.embark.config.storageConfig.upload.provider;
 
-  handleUploadCommand() {
-    const self = this;
-    this.embark.events.setCommandHandler('storage:upload', (cb) => {
-      let platform = this.embark.config.storageConfig.upload.provider;
+  //     let uploadCmds = self.plugins.getPluginsProperty('uploadCmds', 'uploadCmds');
+  //     for (let uploadCmd of uploadCmds) {
+  //       if (uploadCmd.cmd === platform) {
+  //         return uploadCmd.cb.call(uploadCmd.cb, cb);
+  //       }
+  //     }
 
-      let uploadCmds = self.plugins.getPluginsProperty('uploadCmds', 'uploadCmds');
-      for (let uploadCmd of uploadCmds) {
-        if (uploadCmd.cmd === platform) {
-          return uploadCmd.cb.call(uploadCmd.cb, cb);
-        }
-      }
+  //     cb({message: __('platform "{{platform}}" is specified as the upload provider, however no plugins have registered an upload command for "{{platform}}".', {platform: platform})});
+  //   });
+  // }
 
-      cb({message: __('platform "{{platform}}" is specified as the upload provider, however no plugins have registered an upload command for "{{platform}}".', {platform: platform})});
-    });
-  }
+  // addSetProviders(cb) {
+  //   let code = `\nEmbarkJS.Storage.setProviders(${JSON.stringify(this.embark.config.storageConfig.dappConnection || [])}, {web3});`;
 
-  addSetProviders(cb) {
-    let code = `\nEmbarkJS.Storage.setProviders(${JSON.stringify(this.embark.config.storageConfig.dappConnection || [])}, {web3});`;
+  //   let shouldInit = (storageConfig) => {
+  //     return storageConfig.enabled;
+  //   };
 
-    let shouldInit = (storageConfig) => {
-      return storageConfig.enabled;
-    };
+  //   this.embark.addProviderInit('storage', code, shouldInit);
 
-    this.embark.addProviderInit('storage', code, shouldInit);
+  //   async.parallel([
+  //     (next) => {
+  //       if (!this.storageConfig.available_providers.includes('ipfs')) {
+  //         return next();
+  //       }
+  //       this.embark.events.once('ipfs:process:started', next);
+  //     },
+  //     (next) => {
+  //       if (!this.storageConfig.available_providers.includes('swarm')) {
+  //         return next();
+  //       }
+  //       this.embark.events.once('swarm:process:started', next);
+  //     }
+  //   ], (err) => {
+  //     if (err) {
+  //       console.error(__('Error starting storage process(es): %s', err));
+  //     }
 
-    async.parallel([
-      (next) => {
-        if (!this.storageConfig.available_providers.includes('ipfs')) {
-          return next();
-        }
-        this.embark.events.once('ipfs:process:started', next);
-      },
-      (next) => {
-        if (!this.storageConfig.available_providers.includes('swarm')) {
-          return next();
-        }
-        this.embark.events.once('swarm:process:started', next);
-      }
-    ], (err) => {
-      if (err) {
-        console.error(__('Error starting storage process(es): %s', err));
-      }
-
-      this.embark.addConsoleProviderInit('storage', code, shouldInit);
-      // TODO: fix me, this is an ugly workaround for race conditions
-      // in the case where the storage process is too slow when starting up we
-      // execute ourselves the setProviders because the console provider init
-      // was already executed
-      this.embark.events.request('runcode:eval', `if (Object.keys(EmbarkJS.Storage.Providers).length) { ${code} }`, () => {
-        this.ready = true;
-        this.embark.events.emit("module:storage:ready");
-      }, true);
-    });
-  }
+  //     this.embark.addConsoleProviderInit('storage', code, shouldInit);
+  //     // TODO: fix me, this is an ugly workaround for race conditions
+  //     // in the case where the storage process is too slow when starting up we
+  //     // execute ourselves the setProviders because the console provider init
+  //     // was already executed
+  //     this.embark.events.request('runcode:eval', `if (Object.keys(EmbarkJS.Storage.Providers).length) { ${code} }`, () => {
+  //       this.ready = true;
+  //       this.embark.events.emit("module:storage:ready");
+  //     }, true);
+  //   });
+  // }
 
 }
