+++ packages/embark/src/cmd/cmd_controller.js
@@ -169,6 +169,7 @@ class EmbarkController {
         engine.registerModuleGroup("pipeline");
         engine.registerModuleGroup("webserver");
         engine.registerModuleGroup("filewatcher");
+        engine.registerModuleGroup("storage");
 
         engine.events.on('deployment:deployContracts:afterAll', () => {
           console.dir("--- generating files...")
@@ -182,6 +183,10 @@ class EmbarkController {
         plugin.registerActionForEvent("embark:engine:started", async (_params, cb) => {
           await engine.events.request2("blockchain:node:start", engine.config.blockchainConfig, cb);
         });
+        // plugin.registerActionForEvent("embark:engine:started", async (_params, cb) => {
+        //   console.dir("====> requesting storage node to start...")
+        //   await engine.events.request2("storage:node:start", engine.config.storageConfig, cb);
+        // });
 
         engine.events.request('watcher:start');
 
@@ -220,15 +225,18 @@ class EmbarkController {
 
           engine.events.request("webserver:start")
 
-          let contractsFiles = await engine.events.request2("config:contractsFiles");
-          let compiledContracts = await engine.events.request2("compiler:contracts:compile", contractsFiles);
-          let _contractsConfig = await engine.events.request2("config:contractsConfig");
-          let contractsConfig = cloneDeep(_contractsConfig);
-          let [contractsList, contractDependencies] = await engine.events.request2("contracts:build", contractsConfig, compiledContracts);
-          await engine.events.request2("deployment:contracts:deploy", contractsList, contractDependencies);
-          console.dir("deployment done")
+          // await engine.events.request2("storage:node:start", engine.config.storageConfig, () => {});
+          engine.events.request2("storage:node:start", engine.config.storageConfig, () => {});
 
-          await engine.events.request2("watcher:start")
+          // let contractsFiles = await engine.events.request2("config:contractsFiles");
+          // let compiledContracts = await engine.events.request2("compiler:contracts:compile", contractsFiles);
+          // let _contractsConfig = await engine.events.request2("config:contractsConfig");
+          // let contractsConfig = cloneDeep(_contractsConfig);
+          // let [contractsList, contractDependencies] = await engine.events.request2("contracts:build", contractsConfig, compiledContracts);
+          // await engine.events.request2("deployment:contracts:deploy", contractsList, contractDependencies);
+          // console.dir("deployment done")
+//
+          // await engine.events.request2("watcher:start")
         });
       },
       function startDashboard(callback) {
